Add rotary encoder to LiteX SoC

+++ shared/board/fpga_dev_board.py
@@ -46,7 +46,7 @@ class Platform(XilinxPlatform):
     # 29.498MHZ
     default_clk_name = "clk29"
     default_clk_period = 29.498
-    clkFreq = default_clk_period * 1000000
+    clkFreq = (int) (default_clk_period * 1000000)
 
     def __init__(self):
         XilinxPlatform.__init__(self, "xc6slx9-tqg144-2", _io, _connectors)
